feat: update model, add default property, move tools to sheet
