+++ frontend/components/playground/types.ts
@@ -179,6 +179,11 @@ export const providers: { provider: Provider; models: LanguageModel[] }[] = [
       {
         id: "gemini:gemini-2.5-pro-preview-05-06:thinking",
         name: "gemini-2.5-pro-preview-05-06",
+        label: "Gemini 2.5 Pro Preview (05-06)",
+      },
+      {
+        id: "gemini:gemini-2.5-pro-preview-06-05:thinking",
+        name: "gemini-2.5-pro-preview-06-05",
         label: "Gemini 2.5 Pro Preview",
       },
     ],
