+++ frontend/lib/playground/types.ts
@@ -85,4 +85,5 @@ export const googleThinkingModels = [
   "gemini:gemini-2.5-flash-preview-05-20:thinking",
   "gemini:gemini-2.5-pro-exp-03-25:thinking",
   "gemini:gemini-2.5-pro-preview-05-06:thinking",
+  "gemini:gemini-2.5-pro-preview-06-05:thinking",
 ];
